Send data generation functions to separated module
